+++ rust/sglang-server/src/api_server/http/disaggregation/bootstrap.rs
@@ -11,11 +11,11 @@ use arc_swap::ArcSwap;
 use http::{Method, StatusCode};
 use serde::{Deserialize, Serialize};
 
-use crate::api_server::http::{
+use super::super::response::{
     HttpResponse, json_typed_response, method_not_allowed, read_json, text_response,
 };
 use crate::utils::environ;
-use crate::utils::response::json_error;
+use super::super::response::json_error;
 use crate::utils::serialize::{parse_int, parse_int_opt, parse_int_vec};
 
 /// Python default: `SGLANG_DISAGGREGATION_BOOTSTRAP_ENTRY_CLEANUP_INTERVAL`.
