+++ rust/sglang-server/src/utils.rs
@@ -5,7 +5,6 @@ pub mod error;
 pub mod fsm;
 pub mod logging;
 pub mod regex;
-pub mod response;
 pub mod runtime;
 pub mod serialize;
 pub mod sock;
